requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python-headless
numpy
pillow
python-multipart
jinja2